requirements.txt
scipy
numpy
matplotlib
pandas
statsmodels
scikit-learn
pillow
tensorflow
tensorflow-datasets
tensorflow-estimator
tensorflow-gcs-config
tensorflow-hub
tensorflow-metadata
tensorflow-probability
tf-slim
tf_keras
Keras-Preprocessing
tokenizers
torchdata
torchsummary
torchtext
torchvision
tornado
urllib3
scikit-image
seaborn
pytensor
pyarrow
tensorboard
pydot
Sequential
TensorBoard
tensorflow-metal; sys_platform == 'darwin'
